Edit the polynomial class.
